+++ src/__internal__/utils/helpers/guid/index.ts
@@ -7,5 +7,5 @@ export default (): string => {
       .toString(16)
       .substring(1);
   }
-  return `${s4()}${s4()}-${s4()}-${s4()}-${s4()}-${s4()}${s4()}${s4()}`;
+  return `carbon-${s4()}${s4()}-${s4()}-${s4()}-${s4()}-${s4()}${s4()}${s4()}`;
 };